Move classes to new files and prune header includes
Addresses feedback from @KiterLuc
