+++ tiledb/sm/filter/test/CMakeLists.txt
@@ -38,7 +38,15 @@ conclude(unit_test)
 
 commence(unit_test run_filter_pipeline)
     this_target_object_libraries(filter_pipeline)
-    this_target_sources(main.cc unit_run_filter_pipeline.cc)
+    this_target_sources(
+        main.cc
+        add_1_in_place_filter.cc
+        add_1_including_metadata_filter.cc
+        add_1_out_of_place_filter.cc
+        add_n_in_place_filter.cc
+        pseudo_checksum_filter.cc
+        unit_run_filter_pipeline.cc
+    )
 conclude(unit_test)
 
 commence(unit_test float_scale_input_val)
