+++ tiledb/sm/filter/test/unit_run_filter_pipeline.cc
@@ -27,7 +27,11 @@
  *
  * @section DESCRIPTION
  *
- * Tests running the filter pipeline forward and backward.
+ * This set of unit tests checks running the filter pipeline forward and
+ * backward. Most tests here use simplified filters that perform basic
+ * operations such as adding 1 to each value in the original data so that the
+ * filtered data itself can be checked after running the filter pipeline
+ * forward.
  */
 
 #include <random>
@@ -43,10 +47,12 @@
 #include "../encryption_aes256gcm_filter.h"
 #include "../filter.h"
 #include "../filter_pipeline.h"
-#include "../float_scaling_filter.h"
 #include "../positive_delta_filter.h"
-#include "../webp_filter.h"
-#include "../xor_filter.h"
+#include "add_1_in_place_filter.h"
+#include "add_1_including_metadata_filter.h"
+#include "add_1_out_of_place_filter.h"
+#include "add_n_in_place_filter.h"
+#include "pseudo_checksum_filter.h"
 #include "tiledb/common/logger_public.h"
 #include "tiledb/sm/crypto/encryption_key.h"
 #include "tiledb/sm/enums/compressor.h"
@@ -129,474 +135,6 @@ void run_reverse(
                          config)
                      .ok());
 }
-
-/**
- * Simple filter that modifies the input stream by adding 1 to every input
- * element.
- */
-class Add1InPlace : public tiledb::sm::Filter {
- public:
-  // Just use a dummy filter type
-  Add1InPlace(Datatype filter_data_type)
-      : Filter(FilterType::FILTER_NONE, filter_data_type) {
-  }
-
-  void dump(FILE* out) const override {
-    (void)out;
-  }
-
-  Status run_forward(
-      const WriterTile&,
-      WriterTile* const,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output) const override {
-    auto input_size = input->size();
-    RETURN_NOT_OK(output->append_view(input));
-    output->reset_offset();
-
-    uint64_t nelts = input_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t* val = output->value_ptr<uint64_t>();
-      *val += 1;
-      output->advance_offset(sizeof(uint64_t));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-
-    return Status::Ok();
-  }
-
-  Status run_reverse(
-      const Tile&,
-      Tile*,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output,
-      const tiledb::sm::Config& config) const override {
-    (void)config;
-
-    auto input_size = input->size();
-    RETURN_NOT_OK(output->append_view(input));
-    output->reset_offset();
-
-    uint64_t nelts = input_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t* val = output->value_ptr<uint64_t>();
-      *val -= 1;
-      output->advance_offset(sizeof(uint64_t));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-
-    return Status::Ok();
-  }
-
-  Add1InPlace* clone_impl() const override {
-    return tdb_new(Add1InPlace, filter_data_type_);
-  }
-};
-
-/**
- * Simple filter that increments every element of the input stream, writing the
- * output to another buffer. Does not modify the input stream.
- */
-class Add1OutOfPlace : public tiledb::sm::Filter {
- public:
-  // Just use a dummy filter type
-  Add1OutOfPlace(Datatype filter_data_type)
-      : Filter(FilterType::FILTER_NONE, filter_data_type) {
-  }
-
-  void dump(FILE* out) const override {
-    (void)out;
-  }
-
-  Status run_forward(
-      const WriterTile&,
-      WriterTile* const,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output) const override {
-    auto input_size = input->size();
-    auto nelts = input_size / sizeof(uint64_t);
-
-    // Add another output buffer.
-    RETURN_NOT_OK(output->prepend_buffer(input_size));
-    output->reset_offset();
-
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input->read(&inc, sizeof(uint64_t)));
-      inc++;
-      RETURN_NOT_OK(output->write(&inc, sizeof(uint64_t)));
-    }
-
-    // Finish any remaining bytes to ensure no data loss.
-    auto rem = input_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output->write(&byte, sizeof(char)));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-
-    return Status::Ok();
-  }
-
-  Status run_reverse(
-      const Tile&,
-      Tile*,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output,
-      const tiledb::sm::Config& config) const override {
-    (void)config;
-
-    auto input_size = input->size();
-    auto nelts = input->size() / sizeof(uint64_t);
-
-    // Add another output buffer.
-    RETURN_NOT_OK(output->prepend_buffer(input->size()));
-    output->reset_offset();
-
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input->read(&inc, sizeof(uint64_t)));
-      inc--;
-      RETURN_NOT_OK(output->write(&inc, sizeof(uint64_t)));
-    }
-
-    auto rem = input_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output->write(&byte, sizeof(char)));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-
-    return Status::Ok();
-  }
-
-  Add1OutOfPlace* clone_impl() const override {
-    return tdb_new(Add1OutOfPlace, filter_data_type_);
-  }
-};
-
-/**
- * Simple filter that modifies the input stream by adding a constant value to
- * every input element.
- */
-class AddNInPlace : public tiledb::sm::Filter {
- public:
-  // Just use a dummy filter type
-  AddNInPlace(Datatype filter_data_type)
-      : Filter(FilterType::FILTER_NONE, filter_data_type) {
-    increment_ = 1;
-  }
-
-  void dump(FILE* out) const override {
-    (void)out;
-  }
-
-  Status run_forward(
-      const WriterTile&,
-      WriterTile* const,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output) const override {
-    auto input_size = input->size();
-    RETURN_NOT_OK(output->append_view(input));
-    output->reset_offset();
-
-    uint64_t nelts = input_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t* val = output->value_ptr<uint64_t>();
-      *val += increment_;
-      output->advance_offset(sizeof(uint64_t));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-    return Status::Ok();
-  }
-
-  Status run_reverse(
-      const Tile&,
-      Tile*,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output,
-      const tiledb::sm::Config& config) const override {
-    (void)config;
-
-    auto input_size = input->size();
-    RETURN_NOT_OK(output->append_view(input));
-    output->reset_offset();
-
-    uint64_t nelts = input_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t* val = output->value_ptr<uint64_t>();
-      *val -= increment_;
-      output->advance_offset(sizeof(uint64_t));
-    }
-
-    // Metadata not modified by this filter.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-
-    return Status::Ok();
-  }
-
-  uint64_t increment() const {
-    return increment_;
-  }
-
-  void set_increment(uint64_t increment) {
-    increment_ = increment;
-  }
-
-  AddNInPlace* clone_impl() const override {
-    auto clone = tdb_new(AddNInPlace, filter_data_type_);
-    clone->increment_ = increment_;
-    return clone;
-  }
-
- private:
-  uint64_t increment_;
-};
-
-/**
- * Simple filter which computes the sum of its input and prepends the sum
- * to the output. In reverse execute, checks that the sum is correct.
- */
-class PseudoChecksumFilter : public tiledb::sm::Filter {
- public:
-  // Just use a dummy filter type
-  PseudoChecksumFilter(Datatype filter_data_type)
-      : Filter(FilterType::FILTER_NONE, filter_data_type) {
-  }
-
-  void dump(FILE* out) const override {
-    (void)out;
-  }
-
-  Status run_forward(
-      const WriterTile&,
-      WriterTile* const,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output) const override {
-    auto input_size = input->size();
-    auto nelts = input_size / sizeof(uint64_t);
-
-    // The input is unmodified by this filter.
-    RETURN_NOT_OK(output->append_view(input));
-
-    // Forward the existing metadata and prepend a metadata buffer for the
-    // checksum.
-    RETURN_NOT_OK(output_metadata->append_view(input_metadata));
-    RETURN_NOT_OK(output_metadata->prepend_buffer(sizeof(uint64_t)));
-    output_metadata->reset_offset();
-
-    uint64_t sum = 0;
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t val;
-      RETURN_NOT_OK(input->read(&val, sizeof(uint64_t)));
-      sum += val;
-    }
-
-    RETURN_NOT_OK(output_metadata->write(&sum, sizeof(uint64_t)));
-
-    return Status::Ok();
-  }
-
-  Status run_reverse(
-      const Tile&,
-      Tile*,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output,
-      const tiledb::sm::Config& config) const override {
-    (void)config;
-
-    auto input_size = input->size();
-    auto nelts = input_size / sizeof(uint64_t);
-
-    uint64_t input_sum;
-    RETURN_NOT_OK(input_metadata->read(&input_sum, sizeof(uint64_t)));
-
-    uint64_t sum = 0;
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t val;
-      RETURN_NOT_OK(input->read(&val, sizeof(uint64_t)));
-      sum += val;
-    }
-
-    if (sum != input_sum)
-      return Status_FilterError("Filter error; sum does not match.");
-
-    // The output metadata is just a view on the input metadata, skipping the
-    // checksum bytes.
-    RETURN_NOT_OK(output_metadata->append_view(
-        input_metadata,
-        sizeof(uint64_t),
-        input_metadata->size() - sizeof(uint64_t)));
-
-    // The output data is just a view on the unmodified input.
-    RETURN_NOT_OK(output->append_view(input));
-
-    return Status::Ok();
-  }
-
-  PseudoChecksumFilter* clone_impl() const override {
-    return tdb_new(PseudoChecksumFilter, filter_data_type_);
-  }
-};
-
-/**
- * Simple filter that increments every element of the input stream, writing the
- * output to another buffer. The input metadata is treated as a part of the
- * input data.
- */
-class Add1IncludingMetadataFilter : public tiledb::sm::Filter {
- public:
-  // Just use a dummy filter type
-  Add1IncludingMetadataFilter(Datatype filter_data_type)
-      : Filter(FilterType::FILTER_NONE, filter_data_type) {
-  }
-
-  void dump(FILE* out) const override {
-    (void)out;
-  }
-
-  Status run_forward(
-      const WriterTile&,
-      WriterTile* const,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output) const override {
-    auto input_size = static_cast<uint32_t>(input->size()),
-         input_md_size = static_cast<uint32_t>(input_metadata->size());
-    auto nelts = input_size / sizeof(uint64_t),
-         md_nelts = input_md_size / sizeof(uint64_t);
-
-    // Add another output buffer.
-    RETURN_NOT_OK(output->prepend_buffer(input_size + input_md_size));
-    output->reset_offset();
-
-    // Filter input data
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input->read(&inc, sizeof(uint64_t)));
-      inc++;
-      RETURN_NOT_OK(output->write(&inc, sizeof(uint64_t)));
-    }
-    // Finish any remaining bytes to ensure no data loss.
-    auto rem = input_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output->write(&byte, sizeof(char)));
-    }
-
-    // Now filter input metadata.
-    for (uint64_t i = 0; i < md_nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input_metadata->read(&inc, sizeof(uint64_t)));
-      inc++;
-      RETURN_NOT_OK(output->write(&inc, sizeof(uint64_t)));
-    }
-    rem = input_md_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input_metadata->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output->write(&byte, sizeof(char)));
-    }
-
-    // Because this filter modifies the input metadata, we need output metadata
-    // that allows the original metadata to be reconstructed on reverse. Also
-    // note that contrary to most filters, we don't forward the input metadata.
-    RETURN_NOT_OK(output_metadata->prepend_buffer(2 * sizeof(uint32_t)));
-    RETURN_NOT_OK(output_metadata->write(&input_size, sizeof(uint32_t)));
-    RETURN_NOT_OK(output_metadata->write(&input_md_size, sizeof(uint32_t)));
-
-    return Status::Ok();
-  }
-
-  Status run_reverse(
-      const Tile&,
-      Tile*,
-      FilterBuffer* input_metadata,
-      FilterBuffer* input,
-      FilterBuffer* output_metadata,
-      FilterBuffer* output,
-      const tiledb::sm::Config& config) const override {
-    (void)config;
-
-    if (input_metadata->size() != 2 * sizeof(uint32_t))
-      return Status_FilterError("Unexpected input metadata length");
-
-    uint32_t orig_input_size, orig_md_size;
-    RETURN_NOT_OK(input_metadata->read(&orig_input_size, sizeof(uint32_t)));
-    RETURN_NOT_OK(input_metadata->read(&orig_md_size, sizeof(uint32_t)));
-
-    // Add another output buffer.
-    RETURN_NOT_OK(output->prepend_buffer(orig_input_size));
-    // Add another output metadata buffer.
-    RETURN_NOT_OK(output_metadata->prepend_buffer(orig_md_size));
-
-    // Restore original data
-    auto nelts = orig_input_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input->read(&inc, sizeof(uint64_t)));
-      inc--;
-      RETURN_NOT_OK(output->write(&inc, sizeof(uint64_t)));
-    }
-    auto rem = orig_input_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output->write(&byte, sizeof(char)));
-    }
-
-    // Restore original metadata
-    nelts = orig_md_size / sizeof(uint64_t);
-    for (uint64_t i = 0; i < nelts; i++) {
-      uint64_t inc;
-      RETURN_NOT_OK(input->read(&inc, sizeof(uint64_t)));
-      inc--;
-      RETURN_NOT_OK(output_metadata->write(&inc, sizeof(uint64_t)));
-    }
-    rem = orig_md_size % sizeof(uint64_t);
-    for (unsigned i = 0; i < rem; i++) {
-      char byte;
-      RETURN_NOT_OK(input->read(&byte, sizeof(char)));
-      RETURN_NOT_OK(output_metadata->write(&byte, sizeof(char)));
-    }
-
-    return Status::Ok();
-  }
-
-  Add1IncludingMetadataFilter* clone_impl() const override {
-    return tdb_new(Add1IncludingMetadataFilter, filter_data_type_);
-  }
-};
 
 TEST_CASE("Filter: Test empty pipeline", "[filter][empty-pipeline]") {
   tiledb::sm::Config config;
